fix(execution): keep fill ratio typed as double

+++ src/client/quantara_app/lib/features/auto_trade/domain/execution_quality_models.dart
@@ -213,7 +213,7 @@ abstract final class ExecutionQualityMath {
         filledQuantity > plannedQuantity + 1e-9) {
       throw const FormatException('Execution fill quantities are invalid.');
     }
-    return (filledQuantity / plannedQuantity).clamp(0.0, 1.0);
+    return (filledQuantity / plannedQuantity).clamp(0.0, 1.0).toDouble();
   }
 }
